Don't colorize the prompt when in the JetBrains terminal or DISABLE_COLORS is set
The prompt color scheme only works on a black background, but sometimes people
work with a white background. We can't detect the background color, but we can at
least turn colors off in the most likely scenario, which is a JetBrains IDE that
might not be set to the Darcula theme.