add ansible marks to .spacemacs
blockinfile won't allow for multiline regex
because many of the settings are like this:

(thing
  '(
  ...
  ))

Sometimes these are all on one line, sometimes
not if the file owner already put some
settings in there.

We want to put our setting into the middle.
It's more stable to put marks into the default file
for orientation than it is to use a bunch of fancy
replace regexes followed by some blockinfile's.

Example mark in .spacemacs

;; BEGIN ANSIBLE MANAGED section-name
;; END ANSIBLE MANAGED section-name

If this isn't the managed file, the settings will
not be added.