Default to nano rather than vi but allow customization via .zshrc.local
The existing code that tried to configure vi as the standard editor
didn't work, because VISUAL was being set to nano elsewhere. Nano
is probably a better default for machines that will be used by people
of widely varying skillsets, even if most will prefer vi, because it's
self-documenting enough that one can figure out how to edit and save a
file based only on what's on screen.